% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-pairs.R
\name{all_ordered_pairs}
\alias{all_ordered_pairs}
\title{All ordered pairs}
\usage{
all_ordered_pairs(n)
}
\arguments{
\item{n}{Size of underlying set}
}
\value{
A list of all ordered pairs.
}
\description{
All ordered pairs
}
