% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-pairs.R
\name{all_pairs}
\alias{all_pairs}
\title{All unordered pairs}
\usage{
all_pairs(n)
}
\arguments{
\item{n}{Size of underlying set.}
}
\value{
A list of all unordered pairs.
}
\description{
All unordered pairs
}
