% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail.R
\name{avail}
\alias{avail}
\title{Is pair p available in R at cell e?}
\usage{
avail(R, p, e)
}
\arguments{
\item{R}{A partial Room square.}

\item{p}{A pair.}

\item{e}{An empty cell of R.}
}
\value{
True if and only if the pair p can be placed in cell e in R.
}
\description{
Is pair p available in R at cell e?
}
