% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct-pairs.R
\name{distinct_pairs}
\alias{distinct_pairs}
\title{Pairs used in R}
\usage{
distinct_pairs(R)
}
\arguments{
\item{R}{A Room square.}
}
\value{
A list of the distinct pairs that appear in R.
}
\description{
Pairs used in R
}
