% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty-cells.R
\name{empty_cells}
\alias{empty_cells}
\title{Empty cells of a partial Room square}
\usage{
empty_cells(R)
}
\arguments{
\item{R}{A partial Room square.}
}
\value{
A list of empty cells of R.
}
\description{
Empty cells of a partial Room square
}
