% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty-room.R
\name{empty_room}
\alias{empty_room}
\title{Create a partial Room square with no filled cells}
\usage{
empty_room(n = 5)
}
\arguments{
\item{n}{Size of partial Room square to create.}
}
\value{
A partial Room square of size n with no filled cells.
}
\description{
Create a partial Room square with no filled cells
}
