% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-lines.R
\name{grid_lines}
\alias{grid_lines}
\title{Horizontal and vertical grid lines}
\usage{
grid_lines(n_rows, n_cols)
}
\arguments{
\item{n_rows}{Number of rows.}

\item{n_cols}{Number of columns.}
}
\value{
A tibble with columns x, y, xend and yend.
}
\description{
Horizontal and vertical grid lines
}
