% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-lines.R
\name{horiz_lines}
\alias{horiz_lines}
\title{Horizontal grid lines}
\usage{
horiz_lines(n_rows, n_cols)
}
\arguments{
\item{n_rows}{Number of rows.}

\item{n_cols}{Number of columns.}
}
\value{
A tibble with columns x, y, xend and yend.
}
\description{
Horizontal grid lines
}
