% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latin.R
\name{is_col_latin}
\alias{is_col_latin}
\title{Is A Room square column latin?}
\usage{
is_col_latin(R)
}
\arguments{
\item{R}{A Room square}
}
\value{
True if and only if R is column latin.
}
\description{
Is A Room square column latin?
}
