% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latin.R
\name{is_col_latin_i}
\alias{is_col_latin_i}
\title{Does a column satisfy the latin constraint?}
\usage{
is_col_latin_i(R, i)
}
\arguments{
\item{R}{A Room square}

\item{i}{A column index}
}
\value{
True if and only if column i of R satisfies the latin constraint.
}
\description{
Does a column satisfy the latin constraint?
}
