% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room.R
\name{is_maximal_proom}
\alias{is_maximal_proom}
\title{Is R a maximal partial Room square?}
\usage{
is_maximal_proom(R)
}
\arguments{
\item{R}{A partial Room square.}
}
\value{
True if and only if R is a maximal partial Room square, False otherwise.
}
\description{
Is R a maximal partial Room square?
}
