% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room.R
\name{is_partial_room}
\alias{is_partial_room}
\title{Is R a partial Room square?}
\usage{
is_partial_room(R)
}
\arguments{
\item{R}{A partial Room square.}
}
\value{
True if and only if R is a partial Room square, False otherwise.
}
\description{
Is R a partial Room square?
}
