% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room.R
\name{is_room}
\alias{is_room}
\title{Is R a Room square?}
\usage{
is_room(R)
}
\arguments{
\item{R}{A Room square.}
}
\value{
True if and only if R is a Room square, False otherwise.
}
\description{
Is R a Room square?
}
