% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latin.R
\name{is_row_latin}
\alias{is_row_latin}
\title{Is a Room square row latin?}
\usage{
is_row_latin(R)
}
\arguments{
\item{R}{A Room square}
}
\value{
True if and only if R is row latin.
}
\description{
Is a Room square row latin?
}
