% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latin.R
\name{is_row_latin_i}
\alias{is_row_latin_i}
\title{Does a row satisfy the latin constraint?}
\usage{
is_row_latin_i(R, i)
}
\arguments{
\item{R}{A Room square}

\item{i}{A row index}
}
\value{
True if and only if row i of R satisfies the latin constraint.
}
\description{
Does a row satisfy the latin constraint?
}
