% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{n_filled_cells}
\alias{n_filled_cells}
\title{Number of filled cells in a partial Room square}
\usage{
n_filled_cells(R)
}
\arguments{
\item{R}{A partial Room square}
}
\value{
The number of filled cells in R.
}
\description{
Number of filled cells in a partial Room square
}
