% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-room-square.R
\name{plot_room_square}
\alias{plot_room_square}
\title{Plot a Room square}
\usage{
plot_room_square(R)
}
\arguments{
\item{R}{A Room square.}
}
\value{
A ggplot plot of Room square R
}
\description{
Plot a Room square
}
