% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-both.R
\name{remove_both}
\alias{remove_both}
\title{Remove both elements of a pair from a list}
\usage{
remove_both(X, p)
}
\arguments{
\item{X}{A list}

\item{p}{A pair}
}
\value{
The list X with both elements of p removed (if they exist).
}
\description{
Remove both elements of a pair from a list
}
