% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room-class.R
\docType{class}
\name{Room}
\alias{Room}
\title{Create a Room square}
\format{
An \code{\link{R6Class}} generator object
}
\description{
Create a Room square

Create a Room square
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Room-new}{\code{Room$new()}}
\item \href{#method-Room-set}{\code{Room$set()}}
\item \href{#method-Room-is_available}{\code{Room$is_available()}}
\item \href{#method-Room-clone}{\code{Room$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Room-new"></a>}}
\if{latex}{\out{\hypertarget{method-Room-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Room$new(size = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{the order of the Room square to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Room-set"></a>}}
\if{latex}{\out{\hypertarget{method-Room-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Room$set(e, p)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Room-is_available"></a>}}
\if{latex}{\out{\hypertarget{method-Room-is_available}{}}}
\subsection{Method \code{is_available()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Room$is_available(e, p)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Room-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Room-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Room$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
