% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see.R
\name{see}
\alias{see}
\title{Symbols visible from cell e}
\usage{
see(R, e)
}
\arguments{
\item{R}{A Room square.}

\item{e}{A cell in R.}
}
\value{
A list of symbols visible in R from cell e.
}
\description{
Symbols visible from cell e
}
