% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see.R
\name{see2}
\alias{see2}
\title{Symbols visible from cell (row, col) in R}
\usage{
see2(R, row, col)
}
\arguments{
\item{R}{A Room square.}

\item{row}{A row index.}

\item{col}{A column index.}
}
\value{
A list of symbols visible in R from cell (col, rol).
}
\description{
Symbols visible from cell (row, col) in R
}
