% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused-pairs.R
\name{unused_pairs}
\alias{unused_pairs}
\title{Pairs not used in a partial Room square}
\usage{
unused_pairs(R)
}
\arguments{
\item{R}{A partial Room square.}
}
\value{
A list of pairs not used in R.
}
\description{
Pairs not used in a partial Room square
}
