% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-lines.R
\name{vertical_lines}
\alias{vertical_lines}
\title{Vertical grid lines}
\usage{
vertical_lines(n_rows, n_cols)
}
\arguments{
\item{n_rows}{Number of rows.}

\item{n_cols}{Number of columns.}
}
\value{
A tibble with columns x, y, xend and yend.
}
\description{
Vertical grid lines
}
